% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_data.R
\name{agg_data}
\alias{agg_data}
\title{Aggregate data by grouping and time variables}
\usage{
agg_data(x, g_var, d_var, a_unit, ...)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble}.}

\item{g_var}{Grouping variable name.}

\item{d_var}{Date variable. Must be of class \code{Date}.}

\item{a_unit}{Aggregate unit. \code{day}, \code{week}, \code{month} or \code{year}.}

\item{...}{arguments passed to \code{timetk::summarise_by_time} function, including \code{.week_start}.}
}
\value{
a \code{tibble}.
}
\description{
This function aggregates data by grouping and time variables and return time series of same size.
}
\examples{
agg_data(x = dengue_rj, g_var = "ID_MN_RESI", d_var = "DT_SIN_PRI", a_unit = "week")

}
