% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_rj.R
\docType{data}
\name{dengue_rj}
\alias{dengue_rj}
\title{Dengue data on Rio de Janeiro state, Brazil.}
\format{
An object of class \code{data.frame} with 8211 rows and 131 columns.
}
\usage{
dengue_rj
}
\description{
A tibble containing suspected reports of dengue cases for de Rio de Janeiro state, Brazil, for 2020.
}
\keyword{datasets}
