% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_data.R
\name{imp_data}
\alias{imp_data}
\title{This function imputates missing and not valid records from variables using data from pointed best candidate variables.}
\usage{
imp_data(
  x,
  g_var,
  g_var_candidate,
  g_var_nchar,
  d_var,
  d_var_candidate,
  d_var_min,
  d_var_max
)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble}.}

\item{g_var}{character. Grouping variable name.}

\item{g_var_candidate}{character. Variable candidate to imputate \code{g_var}}

\item{g_var_nchar}{numeric. Allowed length for \code{g_var}.}

\item{d_var}{date. Date variable. Must be of class \code{Date}.}

\item{d_var_candidate}{date. Date variable candidate to imputate \code{d_var}.}

\item{d_var_min}{date. Minimum date allowed.}

\item{d_var_max}{date. Maximum date allowed.}
}
\value{
a \code{tibble}.
}
\description{
The records of \code{g_var} variable are checked for presence of missing data (\code{NA}) and for different length than \code{g_var_nchar}. If any of those conditions are met, data from \code{g_var_candidate} is used to imputate \code{g_var} if it is not missing and valid.
}
\details{
The records of \code{d_var} variable are checked for presence of missing data (\code{NA}) and must be in the range of \code{d_var_min} and \code{d_var_max}. If any of those conditions are met, data from \code{d_var_candidate} is used to imputate \code{d_var} if it is not missing and valid.
}
\examples{
imp_data(x = dengue_rj, g_var = "ID_MN_RESI", g_var_candidate = "COMUNINF",
d_var = "DT_SIN_PRI", g_var_nchar = 6, d_var_candidate = "DT_NOTIFIC",
d_var_min = as.Date("2020-01-01"), d_var_max = as.Date("2020-12-31"))

}
